\name{plot.knncat}
\alias{plot.knncat}
\title{Plot a knncat classifier}
\description{Plot a knncat classifier}
\usage{
\method{plot}{knncat}{x, \dots}
}
\arguments{
\item{x}{Knncat object, from \link{knncat}}
}               
\keyword{models}
\details{This plot shows all the estimated numnbers associated with each
level of a variable (or knot, for a continuous variable) in a knncat
classifier.}
\value{None.}
\author{Samuel E. Buttrey, \email{buttrey@nps.edu}}
